refactor(front): inMemoryTestingCacheInstance class
